Add _IO_stdin_used to mpv version script
This symbol is used on some architectures by glibc to determine whether the
calling executable is linked with the old libio ABI or the new libio ABI. All
new executables are supposed to have it defined. Unfortunately, if the version
script does not allow this symbol to be exported, glibc will try to use the
old ABI and cause chaos (crashes in various places).

Gbp-Pq: Name 07_io-stdin-used.patch